palleting fees
